// src/interfaces/topic.ts
type Topic =
  | 'twitch:ban'
  | 'twitch:unban'
  | 'twitch:charity'
  | 'twitch:cheer'
  | 'twitch:follow'
  | 'twitch:goal'
  | 'twitch:hype_train'
  | 'twitch:moderator'
  | 'twitch:poll'
  | 'twitch:prediction'
  | 'twitch:raid'
  | 'twitch:raid_to'
  | 'twitch:redemption'
  | 'twitch:reward'
  | 'twitch:shield_mode'
  | 'twitch:shoutout'
  | 'twitch:shoutout_receive'
  | 'twitch:subscription'
  | 'twitch:update'
  | 'twitch:stream-up'
  | 'twitch:stream-off';

export default Topic;